(#1324) code 画廊图片收费信息科学计数法调整

+++ app/src/main/java/com/zhiyicx/thinksnsplus/modules/dynamic/list/DynamicFragment.java
@@ -403,7 +403,7 @@ public class DynamicFragment extends TSListFragment<DynamicContract.Presenter, D
             imageBean.setImgUrl(task.getImgUrl());
             Toll toll = new Toll();
             toll.setPaid(task.isPaid());
-            toll.setToll_money((float) task.getAmount());
+            toll.setToll_money(task.getAmount());
             toll.setToll_type_string(task.getType());
             toll.setPaid_node(task.getPaid_node());
             imageBean.setToll(toll);
