+++ app/src/main/java/com/zhiyicx/thinksnsplus/modules/personal_center/PersonalCenterFragment.java
@@ -406,7 +406,7 @@ public class PersonalCenterFragment extends TSListFragment<PersonalCenterContrac
             imageBean.setImgUrl(task.getImgUrl());
             Toll toll = new Toll();
             toll.setPaid(task.isPaid());
-            toll.setToll_money((float) task.getAmount());
+            toll.setToll_money(task.getAmount());
             toll.setToll_type_string(task.getType());
             toll.setPaid_node(task.getPaid_node());
             imageBean.setToll(toll);
